# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MÉRITO API — Talent identity & validation platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "409":
          description: Email already in use
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email/password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized

  # ── PROFILES ──────────────────────────────────────────────────────────────
  /profiles/{userId}:
    get:
      operationId: getProfile
      tags: [profiles]
      summary: Get a student profile
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "404":
          description: Not found

    patch:
      operationId: updateProfile
      tags: [profiles]
      summary: Update student profile
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"

  # ── PROJECTS ──────────────────────────────────────────────────────────────
...
  /microprojects:
    get:
      operationId: listMicroprojects
      tags: [microprojects]
      summary: List open micro-project challenges
      parameters:
        - name: skill
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Micro-projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Microproject"

    post:
      operationId: createMicroproject
      tags: [microprojects]
      summary: Post a new micro-project challenge
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MicroprojectInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Microproject"
  /microprojects/{id}:
    get:
      operationId: getMicroproject
      tags: [microprojects]
      summary: Get a micro-project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Micro-project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Microproject"
  /microprojects/{id}/apply:
    post:
      operationId: applyMicroproject
      tags: [microprojects]
      summary: Apply to a micro-project challenge
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationInput"
      responses:
        "201":
          description: Applied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"

  # ── RECRUITER ─────────────────────────────────────────────────────────────
...
  /dashboard/student/{userId}:
    get:
      operationId: getStudentDashboard
      tags: [dashboard]
      summary: Get student dashboard summary — score, project count, skill breakdown, recent activity
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentDashboard"
  /dashboard/score/{userId}:
    get:
      operationId: getScoreBreakdown
      tags: [dashboard]
      summary: Get detailed competence score breakdown
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Score breakdown
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScoreBreakdown"
  /dashboard/activity/{userId}:
    get:
      operationId: getActivityFeed
      tags: [dashboard]
      summary: Get recent activity timeline for a student (returns last 10 items)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Activity feed
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"
...
components:
  schemas:
...
    RegisterInput:
      type: object
      required: [name, email, password, role]
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        role:
          type: string
          enum: [student, recruiter]
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, name, email, role, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [student, recruiter]
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time

    # ── PROFILE SCHEMAS ───────────────────────────────────────────────────
    Profile:
      type: object
      required: [userId, name, bio, score, projectCount, skillCount, validationCount]
      properties:
        userId:
          type: integer
        name:
          type: string
        bio:
          type: ["string", "null"]
        headline:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        githubUrl:
          type: ["string", "null"]
        linkedinUrl:
          type: ["string", "null"]
        portfolioUrl:
          type: ["string", "null"]
        score:
          type: integer
        projectCount:
          type: integer
        skillCount:
          type: integer
        validationCount:
          type: integer
    ProfileUpdate:
      type: object
      properties:
        bio:
          type: string
        headline:
          type: string
        location:
          type: string
        githubUrl:
          type: string
        linkedinUrl:
          type: string
        portfolioUrl:
          type: string

    # ── PROJECT SCHEMAS ───────────────────────────────────────────────────
    Project:
      type: object
      required: [id, userId, title, description, problem, solution, impact, role, technologies, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        description:
          type: string
        problem:
          type: string
        solution:
          type: string
        impact:
          type: string
        role:
          type: string
        technologies:
          type: array
          items:
            type: string
        imageUrl:
          type: ["string", "null"]
        githubUrl:
          type: ["string", "null"]
        figmaUrl:
          type: ["string", "null"]
        demoUrl:
          type: ["string", "null"]
        videoUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...
    Skill:
      type: object
      required: [id, userId, name, category, level]
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        category:
          type: string
          enum: [technical, soft, tool, language]
        level:
          type: string
          enum: [beginner, intermediate, advanced, expert]
...
    Microproject:
      type: object
      required: [id, companyName, title, description, requiredSkills, difficulty, duration, rewardPoints, status, applicantCount, createdAt]
      properties:
        id:
          type: integer
        companyName:
          type: string
        title:
          type: string
        description:
          type: string
        requiredSkills:
          type: array
          items:
            type: string
        difficulty:
          type: string
          enum: [easy, medium, hard]
        duration:
          type: string
        rewardPoints:
          type: integer
        status:
          type: string
          enum: [open, closed, reviewing]
        applicantCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    MicroprojectInput:
      type: object
      required: [title, description, requiredSkills, difficulty, duration, rewardPoints]
      properties:
        title:
          type: string
          minLength: 5
        description:
          type: string
        requiredSkills:
          type: array
          items:
            type: string
        difficulty:
          type: string
          enum: [easy, medium, hard]
        duration:
          type: string
        rewardPoints:
          type: integer
    Application:
      type: object
      required: [id, microprojectId, userId, proposal, status, createdAt]
      properties:
        id:
          type: integer
        microprojectId:
          type: integer
        userId:
          type: integer
        proposal:
          type: string
        status:
          type: string
          enum: [pending, accepted, rejected]
        createdAt:
          type: string
          format: date-time
    ApplicationInput:
      type: object
      required: [proposal]
      properties:
        proposal:
          type: string
          minLength: 20

    # ── RECRUITER SCHEMAS ─────────────────────────────────────────────────
...
    StudentDashboard:
      type: object
      required: [userId, score, projectCount, skillCount, validationCount, sarCount, recentProjects, topSkills]
      properties:
        userId:
          type: integer
        score:
          type: integer
        scoreChange:
          type: integer
        projectCount:
          type: integer
        skillCount:
          type: integer
        validationCount:
          type: integer
        sarCount:
          type: integer
        recentProjects:
          type: array
          items:
            $ref: "#/components/schemas/Project"
        topSkills:
          type: array
          items:
            $ref: "#/components/schemas/Skill"
    ScoreBreakdown:
      type: object
      required: [userId, totalScore, projectsScore, skillsScore, validationsScore, sarScore, activityScore]
      properties:
        userId:
          type: integer
        totalScore:
          type: integer
        projectsScore:
          type: integer
        skillsScore:
          type: integer
        validationsScore:
          type: integer
        sarScore:
          type: integer
        activityScore:
          type: integer
        level:
          type: string
          enum: [emerging, developing, proficient, expert]
        nextMilestone:
          type: integer
    ActivityItem:
      type: object
      required: [id, type, description, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [project_added, skill_added, validation_received, sar_created, microproject_applied]
        description:
          type: string
        createdAt:
          type: string
          format: date-time
...