# lib/api-zod/src/generated/types/microproject.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MÉRITO API — Talent identity & validation platform
 * OpenAPI spec version: 0.1.0
 */
import type { MicroprojectDifficulty } from './microprojectDifficulty';
import type { MicroprojectStatus } from './microprojectStatus';

export interface Microproject {
  id: number;
  companyName: string;
  title: string;
  description: string;
  requiredSkills: string[];
  difficulty: MicroprojectDifficulty;
  duration: string;
  rewardPoints: number;
  status: MicroprojectStatus;
  applicantCount: number;
  createdAt: Date;
}