# lib/api-zod/src/generated/types/microprojectInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MÉRITO API — Talent identity & validation platform
 * OpenAPI spec version: 0.1.0
 */
import type { MicroprojectInputDifficulty } from './microprojectInputDifficulty';

export interface MicroprojectInput {
  /** @minLength 5 */
  title: string;
  description: string;
  requiredSkills: string[];
  difficulty: MicroprojectInputDifficulty;
  duration: string;
  rewardPoints: number;
}