# lib/api-zod/src/generated/types/skill.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MÉRITO API — Talent identity & validation platform
 * OpenAPI spec version: 0.1.0
 */
import type { SkillCategory } from './skillCategory';
import type { SkillLevel } from './skillLevel';

export interface Skill {
  id: number;
  userId: number;
  name: string;
  category: SkillCategory;
  level: SkillLevel;
}