# lib/api-zod/src/generated/types/studentDashboard.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MÉRITO API — Talent identity & validation platform
 * OpenAPI spec version: 0.1.0
 */
import type { Project } from './project';
import type { Skill } from './skill';

export interface StudentDashboard {
  userId: number;
  score: number;
  scoreChange?: number;
  projectCount: number;
  skillCount: number;
  validationCount: number;
  sarCount: number;
  recentProjects: Project[];
  topSkills: Skill[];
}